requirements.txt
# local package
-e .

# external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1

scikit-learn==1.1.1
pandas==1.2.4
numpy==1.20.2
PyYAML==6.0
seaborn==0.11.1
matplotlib==3.4.2
dvc==2.11.0